requirements.txt
#python3.6
Django==3.2.5
psycopg2-binary==2.9.1
gunicorn==20.1.0
djangorestframework==3.14.0